Added a test for passing "primitives" to @Cacheable methods.
This is largely irrelevant as autoboxing means there can't really be primitives flying around, but it does simulate what a user might do and verifies there is no funny business.

This is also more corrent to test here rather than in CacheKeyTests as it uses the same dispatch path, which may have different boxing behaviour.